Basic implementation for plotting and saving
Interaction between argparse and Greengraph added to the entry_point
function, allowing parsing of From, To, Output and Step to the
unmodified Greengraph class. At present this is very rough as all
plotting occurs within the entry_point function. This should be split
into a separate entity to be called or perhaps more appropriately as
function with the Greengraph class itself.